Bugfix - See the commit for more details
This commit fix the following problems:

- Deprecate httpie in favor of cURL because of the issue with big files;
- Fixed the issue where logger was failing for big outputs
- Include the capability to download and upload files inside a different port. Looks like not all Zimbra servers are using 7071 as the SourceAdminPort
- Fixed any issue found by Pycharm
- Fixed an issue with e-mail
- Fixed the log issue "Argument list too long"

+++ installScript/deploy.sh
@@ -5,7 +5,7 @@
 # blacklist_gen: Generate a blacklist of all accounts Zmbackup should ignore
 ################################################################################
 function blacklist_gen(){
-  for ACCOUNT in $(sudo -H -u $OSE_USER bash -c "/opt/zimbra/bin/zmprov -l gaa"); do
+  for ACCOUNT in $(sudo -H -u "$OSE_USER" bash -c "/opt/zimbra/bin/zmprov -l gaa"); do
     if  [[ "$ACCOUNT" = "galsync"* ]] || \
     [[ "$ACCOUNT" = "virus"* ]] || \
     [[ "$ACCOUNT" = "ham"* ]] || \
@@ -14,10 +14,10 @@ function blacklist_gen(){
     [[ "$ACCOUNT" = "zmbackup"* ]] || \
     [[ "$ACCOUNT" = "postmaster"* ]] || \
     [[ "$ACCOUNT" = "root"* ]]; then
-      echo $ACCOUNT >> $ZMBKP_CONF/blacklist.conf
+      echo "$ACCOUNT" >> "$ZMBKP_CONF"/blacklist.conf
     fi
   done
-  echo $ZMBKP_ACCOUNT >> $ZMBKP_CONF/blacklist.conf
+  echo "$ZMBKP_ACCOUNT" >> "$ZMBKP_CONF"/blacklist.conf
 }
 
 ################################################################################
@@ -26,79 +26,77 @@ function blacklist_gen(){
 function deploy_new() {
   echo "Installing... Please wait while we made some changes."
   echo -ne '                      (0%)\r'
-  mkdir -p $OSE_DEFAULT_BKP_DIR > /dev/null 2>&1
-  if [[ $? -ne 0 ]]; then
+  mkdir -p "$OSE_DEFAULT_BKP_DIR" > /dev/null 2>&1
+  BASHERRCODE=$?
+  if [[ $BASHERRCODE -ne 0 ]]; then
         echo "[FAIL] - Can't create the directory"
         echo "For some reason the Zmbackup can't create the folder $OSE_DEFAULT_BKP_DIR."
 	echo "Maybe you are using a NFS and the permissions are wrong?"
 	echo "Please check what happened and try again."
 	uninstall
-	exit $ERR_DEPNOTFOUND
+	exit "$ERR_DEPNOTFOUND"
   fi
 
   if [[ $SESSION_TYPE == "TXT" ]]; then
-    touch $OSE_DEFAULT_BKP_DIR/sessions.txt
+    touch "$OSE_DEFAULT_BKP_DIR"/sessions.txt
   elif [[ $SESSION_TYPE == "SQLITE3" ]]; then
-    sqlite3 $OSE_DEFAULT_BKP_DIR/sessions.sqlite3 < project/lib/sqlite3/database.sql > /dev/null 2>&1
+    sqlite3 "$OSE_DEFAULT_BKP_DIR"/sessions.sqlite3 < project/lib/sqlite3/database.sql > /dev/null 2>&1
   fi
-  chown -R $OSE_USER.$OSE_USER $OSE_DEFAULT_BKP_DIR > /dev/null 2>&1
+  chown -R "$OSE_USER"."$OSE_USER" "$OSE_DEFAULT_BKP_DIR" > /dev/null 2>&1
   echo -ne '#                     (5%)\r'
-  test -d $ZMBKP_CONF || mkdir -p $ZMBKP_CONF
+  test -d "$ZMBKP_CONF" || mkdir -p "$ZMBKP_CONF"
   echo -ne '##                    (10%)\r'
-  test -d $ZMBKP_SRC  || mkdir -p $ZMBKP_SRC
+  test -d "$ZMBKP_SRC"  || mkdir -p "$ZMBKP_SRC"
   echo -ne '###                   (15%)\r'
-  test -d $ZMBKP_SHARE || mkdir -p $ZMBKP_SHARE
-  test -d $ZMBKP_LIB || mkdir -p $ZMBKP_LIB
+  test -d "$ZMBKP_SHARE" || mkdir -p "$ZMBKP_SHARE"
+  test -d "$ZMBKP_LIB" || mkdir -p "$ZMBKP_LIB"
   echo -ne '####                  (20%)\r'
 
   # Disable Parallel's message - Zmbackup remind the user about GNU Parallel
-  mkdir $OSE_INSTALL_DIR/.parallel > /dev/null 2>&1 && touch $OSE_INSTALL_DIR/.parallel/will-cite
-  chown -R zimbra. $OSE_INSTALL_DIR/.parallel
+  mkdir "$OSE_INSTALL_DIR"/.parallel > /dev/null 2>&1 && touch "$OSE_INSTALL_DIR"/.parallel/will-cite
+  chown -R zimbra. "$OSE_INSTALL_DIR"/.parallel
 
   # Copy file
-  install -o $OSE_USER -m 700 $MYDIR/project/zmbackup $ZMBKP_SRC
+  install -o "$OSE_USER" -m 700 "$MYDIR"/project/zmbackup "$ZMBKP_SRC"
   echo -ne '#####                 (25%)\r'
-  cp -R $MYDIR/project/lib/* $ZMBKP_LIB
-  chown -R $OSE_USER. $ZMBKP_LIB
-  chmod -R 700 $ZMBKP_LIB
+  cp -R "$MYDIR"/project/lib/* "$ZMBKP_LIB"
+  chown -R "$OSE_USER". "$ZMBKP_LIB"
+  chmod -R 700 "$ZMBKP_LIB"
   echo -ne '######                (30%)\r'
 
-  # Creating httpie config directory
-  mkdir $OSE_INSTALL_DIR/.httpie > /dev/null 2>&1 && chown zimbra. $OSE_INSTALL_DIR/.httpie
-
-  install --backup=numbered -o root -m 600 $MYDIR/project/config/zmbackup.cron /etc/cron.d/zmbackup
+  install --backup=numbered -o root -m 600 "$MYDIR"/project/config/zmbackup.cron /etc/cron.d/zmbackup
   echo -ne '#######               (35%)\r'
-  install --backup=numbered -o $OSE_USER -m 600 $MYDIR/project/config/zmbackup.conf $ZMBKP_CONF
+  install --backup=numbered -o "$OSE_USER" -m 600 "$MYDIR"/project/config/zmbackup.conf "$ZMBKP_CONF"
   echo -ne '########              (40%)\r'
-  install --backup=numbered -o $OSE_USER -m 600 $MYDIR/project/config/blacklist.conf $ZMBKP_CONF
+  install --backup=numbered -o "$OSE_USER" -m 600 "$MYDIR"/project/config/blacklist.conf "$ZMBKP_CONF"
   echo -ne '#########             (45%)\r'
 
   # Including custom settings
-  sed -i "s|{OSE_DEFAULT_BKP_DIR}|${OSE_DEFAULT_BKP_DIR}|g" $ZMBKP_CONF/zmbackup.conf
+  sed -i "s|{OSE_DEFAULT_BKP_DIR}|${OSE_DEFAULT_BKP_DIR}|g" "$ZMBKP_CONF"/zmbackup.conf
   echo -ne '##########            (50%)\r'
-  sed -i "s|{ZMBKP_ACCOUNT}|${ZMBKP_ACCOUNT}|g" $ZMBKP_CONF/zmbackup.conf
+  sed -i "s|{ZMBKP_ACCOUNT}|${ZMBKP_ACCOUNT}|g" "$ZMBKP_CONF"/zmbackup.conf
   echo -ne '###########           (55%)\r'
-  sed -i "s|{ZMBKP_PASSWORD}|${ZMBKP_PASSWORD}|g" $ZMBKP_CONF/zmbackup.conf
+  sed -i "s|{ZMBKP_PASSWORD}|${ZMBKP_PASSWORD}|g" "$ZMBKP_CONF"/zmbackup.conf
   echo -ne '############          (60%)\r'
-  sed -i "s|{ZMBKP_MAIL_ALERT}|${ZMBKP_MAIL_ALERT}|g" $ZMBKP_CONF/zmbackup.conf
+  sed -i "s|{ZMBKP_MAIL_ALERT}|${ZMBKP_MAIL_ALERT}|g" "$ZMBKP_CONF"/zmbackup.conf
   echo -ne '#############         (65%)\r'
-  sed -i "s|{ZMBKP_MAIL_SENDER}|${ZMBKP_MAIL_SENDER}|g" $ZMBKP_CONF/zmbackup.conf
+  sed -i "s|{ZMBKP_MAIL_SENDER}|${ZMBKP_MAIL_SENDER}|g" "$ZMBKP_CONF"/zmbackup.conf
   echo -ne '#############         (65%)\r'
-  sed -i "s|{OSE_INSTALL_ADDRESS}|${OSE_INSTALL_ADDRESS}|g" $ZMBKP_CONF/zmbackup.conf
-  sed -i "s|{ZMBKP_MAIL_SENDER}|${OSE_INSTALL_ADDRESS}|g" $ZMBKP_CONF/zmbackup.conf
+  sed -i "s|{OSE_INSTALL_ADDRESS}|${OSE_INSTALL_ADDRESS}|g" "$ZMBKP_CONF"/zmbackup.conf
+  sed -i "s|{ZMBKP_MAIL_SENDER}|${OSE_INSTALL_ADDRESS}|g" "$ZMBKP_CONF"/zmbackup.conf
   echo -ne '##############        (70%)\r'
-  sed -i "s|{OSE_INSTALL_LDAPPASS}|${OSE_INSTALL_LDAPPASS}|g" $ZMBKP_CONF/zmbackup.conf
-  sed -i "s|{SESSION_TYPE}|${SESSION_TYPE}|g" $ZMBKP_CONF/zmbackup.conf
+  sed -i "s|{OSE_INSTALL_LDAPPASS}|${OSE_INSTALL_LDAPPASS}|g" "$ZMBKP_CONF"/zmbackup.conf
+  sed -i "s|{SESSION_TYPE}|${SESSION_TYPE}|g" "$ZMBKP_CONF"/zmbackup.conf
   echo -ne '###############       (75%)\r'
-  sed -i "s|{OSE_USER}|${OSE_USER}|g" $ZMBKP_CONF/zmbackup.conf
-  sed -i "s|{MAX_PARALLEL_PROCESS}|${MAX_PARALLEL_PROCESS}|g" $ZMBKP_CONF/zmbackup.conf
+  sed -i "s|{OSE_USER}|${OSE_USER}|g" "$ZMBKP_CONF"/zmbackup.conf
+  sed -i "s|{MAX_PARALLEL_PROCESS}|${MAX_PARALLEL_PROCESS}|g" "$ZMBKP_CONF"/zmbackup.conf
   echo -ne '################      (80%)\r'
-  sed -i "s|{ROTATE_TIME}|${ROTATE_TIME}|g" $ZMBKP_CONF/zmbackup.conf
-  sed -i "s|{LOCK_BACKUP}|${LOCK_BACKUP}|g" $ZMBKP_CONF/zmbackup.conf
+  sed -i "s|{ROTATE_TIME}|${ROTATE_TIME}|g" "$ZMBKP_CONF"/zmbackup.conf
+  sed -i "s|{LOCK_BACKUP}|${LOCK_BACKUP}|g" "$ZMBKP_CONF"/zmbackup.conf
   echo -ne '#################     (85%)\r'
 
   # Fix backup dir permissions (owner MUST be $OSE_USER)
-  chown $OSE_USER $OSE_DEFAULT_BKP_DIR
+  chown "$OSE_USER" "$OSE_DEFAULT_BKP_DIR"
   echo -ne '##################    (90%)\r'
 
   # Generate Zmbackup's blacklist
@@ -106,14 +104,15 @@ function deploy_new() {
   echo -ne '###################   (95%)\r'
 
   # Creating Zmbackup backup user
-  STATUS=$((sudo -H -u $OSE_USER bash -c "/opt/zimbra/bin/zmprov ca '$ZMBKP_ACCOUNT' '$ZMBKP_PASSWORD' zimbraIsAdminAccount TRUE zimbraAdminAuthTokenLifetime 1") 2>&1)
-  if [[ $? -eq 0 ]]; then
+  STATUS=$( (sudo -H -u "$OSE_USER" bash -c "/opt/zimbra/bin/zmprov ca '$ZMBKP_ACCOUNT' '$ZMBKP_PASSWORD' zimbraIsAdminAccount TRUE zimbraAdminAuthTokenLifetime 1") 2>&1)
+  BASHERRCODE=$?
+  if [[ $BASHERRCODE -eq 0 ]]; then
       echo -ne '####################  (100%)\r'
   else
     echo "ERROR - Can't create the user. Executing rollback process"
     echo "Error description: $STATUS"
     uninstall
-    exit $ERR_CREATE_USER
+    exit "$ERR_CREATE_USER"
   fi
 }
 
@@ -124,23 +123,20 @@ function deploy_upgrade(){
   # Removing old version
   echo "Upgrading... Please wait while we made some changes."
   echo -ne '                     (0%)\r'
-  rm -rf $ZMBKP_SHARE $ZMBKP_SRC/zmbhousekeep > /dev/null 2>&1
+  rm -rf "$ZMBKP_SHARE" "$ZMBKP_SRC"/zmbhousekeep > /dev/null 2>&1
   echo -ne '##########            (50%)\r'
 
   # Disable Parallel's message - Zmbackup remind the user about GNU Parallel
-  mkdir $OSE_INSTALL_DIR/.parallel > /dev/null 2>&1 && touch $OSE_INSTALL_DIR/.parallel/will-cite
-  chown -R zimbra. $OSE_INSTALL_DIR/.parallel
-
-  # Creating httpie config directory
-  mkdir $OSE_INSTALL_DIR/.httpie > /dev/null 2>&1 && chown zimbra. $OSE_INSTALL_DIR/.httpie
+  mkdir "$OSE_INSTALL_DIR"/.parallel > /dev/null 2>&1 && touch "$OSE_INSTALL_DIR"/.parallel/will-cite
+  chown -R zimbra. "$OSE_INSTALL_DIR"/.parallel
 
   # Copy files
-  install -o $OSE_USER -m 700 $MYDIR/project/zmbackup $ZMBKP_SRC
+  install -o "$OSE_USER" -m 700 "$MYDIR"/project/zmbackup "$ZMBKP_SRC"
   echo -ne '###############       (75%)\r'
-  test -d $ZMBKP_LIB || mkdir -p $ZMBKP_LIB
-  cp -R $MYDIR/project/lib/* $ZMBKP_LIB
-  chown -R $OSE_USER. $ZMBKP_LIB
-  chmod -R 700 $ZMBKP_LIB
+  test -d "$ZMBKP_LIB" || mkdir -p "$ZMBKP_LIB"
+  cp -R "$MYDIR"/project/lib/* "$ZMBKP_LIB"
+  chown -R "$OSE_USER". "$ZMBKP_LIB"
+  chmod -R 700 "$ZMBKP_LIB"
   echo -ne '####################  (100%)\r'
 }
 
@@ -149,27 +145,26 @@ function deploy_upgrade(){
 ################################################################################
 function uninstall() {
   echo "Removing... Please wait while we made some changes."
-  source $ZMBKP_CONF/zmbackup.conf
+  source "$ZMBKP_CONF"/zmbackup.conf
   echo -ne '                     (0%)\r'
-  rm -rf $ZMBKP_SHARE $ZMBKP_SRC/zmbhousekeep > /dev/null 2>&1
-  rm -rf $OSE_INSTALL_DIR/.parallel
-  rm -rf $OSE_INSTALL_DIR/.httpie
+  rm -rf "$ZMBKP_SHARE" "$ZMBKP_SRC"/zmbhousekeep > /dev/null 2>&1
+  rm -rf "$OSE_INSTALL_DIR"/.parallel
   echo -ne '#####                 (25%)\r'
   rm -rf /etc/yum.repos.d/tange.repo
   rm -rf /etc/cron.d/zmbackup
-  rm -rf $ZMBKP_LIB $ZMBKP_CONF $ZMBKP_SRC/zmbackup
+  rm -rf "$ZMBKP_LIB" "$ZMBKP_CONF" "$ZMBKP_SRC"/zmbackup
   echo -ne '##########            (50%)\r'
   if [[ -f $ZMBKP_CONF/blacklist.conf ]]; then
-    install --backup=numbered -o $OSE_USER -m 600 $MYDIR/project/config/blacklist.conf $ZMBKP_CONF
+    install --backup=numbered -o "$OSE_USER" -m 600 "$MYDIR"/project/config/blacklist.conf "$ZMBKP_CONF"
     blacklist_gen
   fi
   echo -ne '###############       (75%)\r'
-  sudo -H -u $OSE_USER bash -c "/opt/zimbra/bin/zmprov da $ZMBKP_ACCOUNT" > /dev/null 2>&1
+  sudo -H -u "$OSE_USER" bash -c "/opt/zimbra/bin/zmprov da $ZMBKP_ACCOUNT" > /dev/null 2>&1
   echo -ne '####################  (100%)\r'
   printf "Preserve Backup Storage?[n/Y]"
-  read OPT
+  read -r OPT
   if [[ $OPT == 'N' && $OPT == 'n' ]]; then
     echo "Removing backup storage..."
-    rm -rf $WORKDIR/*
+    rm -rf "${WORKDIR:?}"/*
   fi
 }
